added mocks module for e2e test
